feat: improvements to travel plan controller

+++ src/main/java/com/stuflo/roamiospring/controllers/TravelPlanController.java
@@ -1,10 +1,12 @@
 package com.stuflo.roamiospring.controllers;
 
 import com.stuflo.roamiospring.dtos.TravelPlanDto;
+import com.stuflo.roamiospring.models.User;
 import com.stuflo.roamiospring.responses.TravelPlanResponse;
 import com.stuflo.roamiospring.services.TravelPlanService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,15 +21,15 @@ public class TravelPlanController {
     }
 
     @GetMapping
-    ResponseEntity<List<TravelPlanResponse>> readTravelPlan() {
-        List<TravelPlanResponse> travelPlans = service.findAll();
+    ResponseEntity<List<TravelPlanResponse>> readTravelPlan(@AuthenticationPrincipal User user) {
+        List<TravelPlanResponse> travelPlans = service.findAll(user.getId());
 
         return ResponseEntity.ok(travelPlans);
     }
 
     @GetMapping("/{id}")
-    ResponseEntity<TravelPlanResponse> readTravelPlanById(@PathVariable Long id) {
-        TravelPlanResponse travelPlan = service.findById(id);
+    ResponseEntity<TravelPlanResponse> readTravelPlanById(@PathVariable Long id, @AuthenticationPrincipal User user) {
+        TravelPlanResponse travelPlan = service.findById(user.getId(), id);
 
         if (travelPlan == null) {
             return ResponseEntity.notFound().build();
@@ -37,9 +39,9 @@ public class TravelPlanController {
     }
 
     @PostMapping
-    ResponseEntity<TravelPlanResponse> createTravelPlan(@RequestBody TravelPlanDto travelPlan) {
+    ResponseEntity<TravelPlanResponse> createTravelPlan(@RequestBody TravelPlanDto travelPlan, @AuthenticationPrincipal User user) {
         try {
-            TravelPlanResponse createdPlan = service.createTravelPlan(travelPlan);
+            TravelPlanResponse createdPlan = service.createTravelPlan(travelPlan, user.getId());
 
             return ResponseEntity.status(HttpStatus.CREATED).body(createdPlan);
         } catch (Exception e) {
@@ -48,14 +50,24 @@ public class TravelPlanController {
     }
 
     @PutMapping("/{id}")
-    ResponseEntity<TravelPlanResponse> updateTravelPlan(@RequestBody TravelPlanDto travelPlan, @PathVariable Long id) {
-        TravelPlanResponse updatedPlan = service.updateTravelPlan(travelPlan, id);
+    ResponseEntity<TravelPlanResponse> updateTravelPlan(@RequestBody TravelPlanDto travelPlan, @PathVariable Long id, @AuthenticationPrincipal User user) {
+        TravelPlanResponse updatedPlan = service.updateTravelPlan(travelPlan, user.getId(), id);
+
+        if (updatedPlan == null) {
+            return ResponseEntity.notFound().build();
+        }
 
         return ResponseEntity.ok(updatedPlan);
     }
 
     @DeleteMapping("/{id}")
-    ResponseEntity<Void> deleteTravelPlan(@PathVariable Long id) {
+    ResponseEntity<Void> deleteTravelPlan(@PathVariable Long id, @AuthenticationPrincipal User user) {
+        TravelPlanResponse tp = service.findById(user.getId(), id);
+
+        if (tp == null) {
+            return ResponseEntity.notFound().build();
+        }
+
         service.deleteById(id);
 
         return ResponseEntity.ok().build();
